# Fewer prove/verify allocations