The changes in this PR were driven by heap profiling applied to a single iteration of the Varuna `snark_verify` benchmark; they are mostly quite straightforward and provide the following benefits:
```
allocations:    13,504 -> 12,237 (-9.4%)
reallocations:  1,979 -> 712 (-64%)
max memory use: 14.13 MiB -> 13.70 MiB (-3%)
runtime:        202.3 ms -> 196.8 ms (-2.7%)
runtime σ:      3.0 ms -> 1.4 ms (-53.3%)
```
(the runtime was measured with `hyperfine`)